MIMM/requirements.txt
absl-py==0.14.1
adjustText==0.8
aiohttp==3.7.4.post0
albumentations==1.3.0
altair==4.2.2
antlr4-python3-runtime==4.8
appdirs==1.4.4
argcomplete==1.12.3
argon2-cffi==21.1.0
asciitree==0.3.3
async-timeout==3.0.1
attrs==21.2.0
autopep8==2.0.4
backcall==0.2.0
backports.zoneinfo==0.2.1
batchgenerators==0.24
bioinfokit==2.1.0
bleach==4.1.0
blinker==1.5
bnlearn==0.6.3
cachetools==4.2.4
cairocffi==1.3.0
CairoSVG==2.5.2
causal-learn==0.1.2.8
causality==0.0.10
celluloid==0.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
community==1.0.0b1
ConfigArgParse==1.5.3
configparser==5.2.0
coverage==6.4.2
cssselect2==0.6.0
cycler==0.10.0
data==0.4
dcor==0.5.3
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
df2onehot==1.0.1
dicom2nifti==2.3.0
dill==0.3.6
docker-pycreds==0.4.0
dowhy==0.8
entrypoints==0.3
fasteners==0.18
filelock==3.13.1
Flask==2.1.3
fsspec==2021.10.1
funcsigs==1.0.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.0
google-auth-oauthlib==0.4.6
grad-cam==1.4.2
graphviz==0.19.1
grpcio==1.41.0
h5py==3.8.0
humanize==3.12.0
hydra-core==1.1.1
idna==3.3
idx2numpy==1.2.3
imageio==2.9.0
imbalanced-learn==0.9.0
imblearn==0.0
imgaug==0.4.0
importlib-metadata==4.8.1
importlib-resources==5.4.0
iniconfig==1.1.1
install==1.3.5
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
ismember==1.0.1
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.0.2
joblib==1.1.0
jsonpickle==2.2.0
jsonschema==4.1.0
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kaleido==0.2.1
kiwisolver==1.3.2
latex==0.7.0
linecache2==1.0.0
llvmlite==0.39.1
Markdown==3.3.4
markdown-it-py==2.2.0
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
matplotlib-venn==0.11.9
mdurl==0.1.2
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==5.2.0
multiprocess==0.70.14
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
nibabel==3.2.1
notebook==6.4.4
numba==0.56.3
numcodecs==0.10.2
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.1
olefile==0.46
omegaconf==2.1.1
opencv-python==4.5.5.64
opencv-python-headless==4.7.0.72
p-tqdm==1.4.0
packaging==23.2
panda==0.3.1
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.2
pathos==0.3.0
pathtools==0.1.2
patsy==0.5.2
pexpect==4.8.0
pgmpy==0.1.19
pickleshare==0.7.5
Pillow==8.4.0
plotly==5.13.1
pluggy==1.0.0
pox==0.3.2
ppft==1.7.6.6
prettytable==2.5.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.11.1
pycore==17.4.17
pycparser==2.20
pydeck==0.8.0
pyDeprecate==0.3.1
pydicom==2.2.2
pydot==1.4.2
Pygments==2.16.1
Pympler==1.0.1
pyparsing==2.4.7
pypickle==1.1.0
pyrsistent==0.18.0
pytest==7.1.2
pytest-cov==3.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytorch-lightning==1.4.9
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyvis==0.2.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
qudida==0.0.4
requests==2.26.0
requests-oauthlib==1.3.0
rich==13.3.2
rope==1.11.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
sentry-sdk==1.5.2
setproctitle==1.2.2
Shapely==1.7.1
shortuuid==1.0.8
shutilwhich==1.1.0
SimpleITK==2.1.1
simplejson==3.17.6
six==1.16.0
sklearn==0.0
slurm==0.4.6
smmap==5.0.0
spark-parser==1.8.9
statsmodels==0.13.2
streamlit==1.20.0
subprocess32==3.5.4
sympy==1.10.1
tabulate==0.8.10
tempdir==0.7.1
tenacity==8.2.2
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
textwrap3==0.9.2
threadpoolctl==3.0.0
tifffile==2021.10.12
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.10.0
torchaudio==0.10.0
torchinfo==1.8.0
torchio==0.18.57
torchmetrics==0.5.1
torchsummary==1.5.1
torchvision==0.11.1
torchviz==0.0.2
tornado==6.1
tqdm==4.62.3
traceback2==1.4.0
traitlets==5.1.0
triton==2.1.0
ttach==0.0.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.3
uncompyle6==3.8.0
unittest2==1.1.0
urllib3==1.26.7
utils==1.0.1
validators==0.20.0
vtk==9.2.6
wandb==0.12.11
watchdog==3.0.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.13.2
xdis==6.0.3
yarl==1.7.0
yaspin==2.1.0
zarr==2.12.0
zipp==3.6.0
